Bug fixes and add schedule print util
Storing gp at a stack offset of 0 writes it outside of the stack.
Function definitions in header files should be static, or else you get a problem with multiple definitions if you include the header in multiple files.
We can't use caps stored at index 0 of the cap table for syscalls, because an argument of 0 is assumed to be no cap.