Bug fix and various fixes
The bug fix regards rarely failing to return true from a yield. The other fixes are mostly sanity updates that ensure preemption is enabled only when it should and that we don't overwrite other bits than the preemption bit in mstatus. Also a slight cleanup in sn_recv_delete_ts